backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2023.11.17
    # via
    #   -r requirements.in
    #   cloudinary
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
cloudinary==1.37.0
    # via
    #   -r requirements.in
    #   django-cloudinary-storage
contourpy==1.2.0
    # via
    #   -r requirements.in
    #   matplotlib
coverage==7.4.0
    # via -r requirements-dev.in
cycler==0.12.1
    # via
    #   -r requirements.in
    #   matplotlib
distlib==0.3.8
    # via -r requirements.in
django==4.1.13
    # via
    #   -r requirements.in
    #   django-ckeditor
    #   django-cors-headers
    #   django-js-asset
    #   djangorestframework
    #   djangorestframework-simplejwt
django-ckeditor==6.7.0
    # via -r requirements.in
django-cleanup==8.0.0
    # via -r requirements.in
django-cloudinary-storage==0.3.0
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-js-asset==2.2.0
    # via django-ckeditor
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-jwt==1.11.0
    # via -r requirements.in
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
environ==1.0
    # via -r requirements.in
filelock==3.13.1
    # via -r requirements.in
fonttools==4.47.0
    # via
    #   -r requirements.in
    #   matplotlib
idna==3.6
    # via
    #   -r requirements.in
    #   requests
kiwisolver==1.4.5
    # via
    #   -r requirements.in
    #   matplotlib
markdown2==2.4.12
    # via -r requirements.in
matplotlib==3.8.2
    # via -r requirements.in
numpy==1.26.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   opencv-contrib-python
opencv-contrib-python==4.8.1.78
    # via -r requirements.in
packaging==23.2
    # via
    #   -r requirements.in
    #   matplotlib
pillow==10.1.0
    # via
    #   -r requirements.in
    #   matplotlib
platformdirs==4.1.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyjwt==1.7.1
    # via
    #   -r requirements.in
    #   djangorestframework-jwt
    #   djangorestframework-simplejwt
pyparsing==3.1.1
    # via
    #   -r requirements.in
    #   matplotlib
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   matplotlib
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-cloudinary-storage
six==1.16.0
    # via
    #   -r requirements.in
    #   cloudinary
    #   python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via asgiref
tzdata==2023.3
    # via django
urllib3==2.1.0
    # via
    #   -r requirements.in
    #   cloudinary
    #   requests
